+++ backend/tests/Unit/Middleware/Psr15/RateLimitIPTest.php
@@ -6,7 +6,7 @@ declare(strict_types=1);
 namespace Tests\Unit\Middleware\Psr15;
 
 use Neucore\Middleware\Psr15\RateLimit;
-use Neucore\Middleware\Psr15\RateLimitGlobal;
+use Neucore\Middleware\Psr15\RateLimitIP;
 use Neucore\Service\Config;
 use Neucore\Storage\ApcuStorage;
 use Neucore\Storage\Variables;
@@ -16,7 +16,7 @@ use Tests\Logger;
 use Tests\RequestFactory;
 use Tests\RequestHandler;
 
-class RateLimitGlobalTest extends TestCase
+class RateLimitIPTest extends TestCase
 {
     private static string $ip = '127.0.0.1';
 
@@ -26,18 +26,18 @@ class RateLimitGlobalTest extends TestCase
 
     public static function setUpBeforeClass(): void
     {
-        RateLimitGlobal::$active = true;
+        RateLimitIP::$active = true;
 
         $_SERVER['REMOTE_ADDR'] = self::$ip;
         $_SERVER['HTTP_AUTHORIZATION'] = 'Bearer ' . base64_encode(self::$appIdp.':abc');
-        self::$key = Variables::GLOBAL_RATE_LIMIT . '_' . str_replace('.', '', self::$ip);
+        self::$key = Variables::RATE_LIMIT_IP . '_' . str_replace('.', '', self::$ip);
     }
 
     public static function tearDownAfterClass(): void
     {
         unset($_SERVER['REMOTE_ADDR']);
         unset($_SERVER['HTTP_AUTHORIZATION']);
-        RateLimitGlobal::$active = false;
+        RateLimitIP::$active = false;
     }
 
     public function testProcess_Inactive()
@@ -49,17 +49,17 @@ class RateLimitGlobalTest extends TestCase
         $request = RequestFactory::createRequest();
 
         $config1 = new Config(['rate_limit' => ['max' => 0, 'time' => 0]]);
-        $middleware1 = new RateLimitGlobal($storage, $config1, new ResponseFactory(), $logger);
+        $middleware1 = new RateLimitIP($storage, $config1, new ResponseFactory(), $logger);
         $response1 = $middleware1->process($request, new RequestHandler());
         $this->assertSame(200, $response1->getStatusCode());
 
         $config2 = new Config(['rate_limit' => ['max' => 50, 'time' => 0]]);
-        $middleware2 = new RateLimitGlobal($storage, $config2, new ResponseFactory(), $logger);
+        $middleware2 = new RateLimitIP($storage, $config2, new ResponseFactory(), $logger);
         $response2 = $middleware2->process($request, new RequestHandler());
         $this->assertSame(200, $response2->getStatusCode());
 
         $config3 = new Config(['rate_limit' => ['max' => 0, 'time' => 10]]);
-        $middleware3 = new RateLimitGlobal($storage, $config3, new ResponseFactory(), $logger);
+        $middleware3 = new RateLimitIP($storage, $config3, new ResponseFactory(), $logger);
         $response3 = $middleware3->process($request, new RequestHandler());
         $this->assertSame(200, $response3->getStatusCode());
     }
@@ -73,21 +73,21 @@ class RateLimitGlobalTest extends TestCase
         $request = RequestFactory::createRequest();
 
         $config = new Config(['rate_limit' => ['max' => 50, 'time' => 10]]);
-        $middleware = new RateLimitGlobal($storage, $config, new ResponseFactory(), $logger);
+        $middleware = new RateLimitIP($storage, $config, new ResponseFactory(), $logger);
         $response = $middleware->process($request, new RequestHandler());
         $this->assertSame(429, $response->getStatusCode());
 
         $this->assertSame('-1', $response->getHeader(RateLimit::HEADER_REMAIN)[0]);
         $this->assertEqualsWithDelta(4.5, $response->getHeader(RateLimit::HEADER_RESET)[0], 1.0);
         $this->assertStringStartsWith(
-            'Global rate limit exceeded with 51 requests in ', // ... ~5.5 seconds
+            'IP rate limit exceeded with 51 requests in ', // ... ~5.5 seconds
             $response->getBody()->__toString()
         );
 
         $logs = $logger->getHandler()->getRecords();
         $this->assertSame(1, count($logs));
         $this->assertStringStartsWith(
-            'Global Rate Limit: IP '.self::$ip.', App-ID '.self::$appIdp.', '.
+            'IP Rate Limit: '.self::$ip.', App-ID '.self::$appIdp.', '.
                 'limit exceeded with 51 request in ', // ... ~5.5 seconds.
             $logs[0]['message']
         );